Moved Planechase portion of content to its own component.
